+++ app/views/EditSourceDatasetView/common/entities.ts
@@ -1,0 +1,4 @@
+import { InferType } from "yup";
+import { sourceDatasetEditSchema } from "./schema";
+
+export type SourceDatasetEditData = InferType<typeof sourceDatasetEditSchema>;
